Resolving issues
