+++ tsconfig.json
@@ -25,5 +25,4 @@
 	},
 	"include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "test.js"],
 	"exclude": ["node_modules"]
-
 }
